[POP] Refactor - group authentication scheme files
