+++ src/client/Microsoft.Identity.Client/AuthScheme/Bearer/BearerAuthenticationScheme.cs
@@ -4,13 +4,15 @@
 using System.Collections.Generic;
 using Microsoft.Identity.Client.Cache.Items;
 
-namespace Microsoft.Identity.Client
+namespace Microsoft.Identity.Client.AuthScheme.Bearer
 {
     internal class BearerAuthenticationScheme : IAuthenticationScheme
     {
+        internal const string BearerTokenType = "bearer";
+
         public string AuthorizationHeaderPrefix => "Bearer";
 
-        public string ExpectedTokenType => "bearer";
+        public string AccessTokenType => BearerTokenType;
 
         public string KeyId => null;
 
@@ -21,7 +23,7 @@ namespace Microsoft.Identity.Client
 
         public IDictionary<string, string> GetTokenRequestParams()
         {
-            // ESTS issues bearer tokens by default, no need for any extra params
+            // ESTS issues Bearer tokens by default, no need for any extra params
             return new Dictionary<string, string>();
         }
     }
