+++ src/client/Microsoft.Identity.Client/AuthScheme/SSHCertificates/SSHCertAuthenticationScheme.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using Microsoft.Identity.Client.Cache.Items;
 using Microsoft.Identity.Client.OAuth2;
 
-namespace Microsoft.Identity.Client.SSHCertificates
+namespace Microsoft.Identity.Client.AuthScheme.SSHCertificates
 {
     internal class SSHCertAuthenticationScheme : IAuthenticationScheme
     {
@@ -33,7 +33,7 @@ namespace Microsoft.Identity.Client.SSHCertificates
             throw new MsalClientException(
                 MsalError.SSHCertUsedAsHttpHeader,
                 MsalErrorMessage.SSHCertUsedAsHttpHeader);
-        public string ExpectedTokenType => SSHCertTokenType;
+        public string AccessTokenType => SSHCertTokenType;
 
 
         public string KeyId { get; }
