+++ src/client/Microsoft.Identity.Client/AuthenticationResult.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Globalization;
+using Microsoft.Identity.Client.AuthScheme;
 using Microsoft.Identity.Client.Cache.Items;
 
 namespace Microsoft.Identity.Client
