+++ src/client/Microsoft.Identity.Client/Platforms/net45/NetDesktopPoPCryptoProvider.cs
@@ -3,7 +3,7 @@
 
 using System;
 using System.Security.Cryptography;
-using Microsoft.Identity.Client.PoP;
+using Microsoft.Identity.Client.AuthScheme.PoP;
 using Microsoft.Identity.Client.Utils;
 
 namespace Microsoft.Identity.Client.Platforms.net45
